--- resources/blocks/index.ts
@@ -1,4 +1,0 @@
-export * from './FlightsCard';
-export * from './MapCard';
-export * from './ProfileCard';
-export * from './StatsCard';
