+++ resources/pages/Flights/Flights.tsx
@@ -1,3 +1,3 @@
-import { FlightsCard } from '../../blocks';
+import { FlightsCard } from './FlightsCard';
 
 export const Flights = (): JSX.Element => <FlightsCard />;
