+++ resources/pages/Profile/Profile.tsx
@@ -1,4 +1,5 @@
-import { MapCard, ProfileCard } from '../../blocks';
+import { MapCard } from './MapCard';
+import { ProfileCard } from './ProfileCard';
 import { DashboardContainer } from '../../common/components';
 
 export const Profile = (): JSX.Element => (
